% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeByMonth.R
\name{SummarizeByMonth}
\alias{SummarizeByMonth}
\title{Summarize data frame by Month
This function summarizes all columns by week and returns a dataframe with summarized data and month dates}
\usage{
SummarizeByMonth(df, summarizationMethod, flipNA)
}
\arguments{
\item{df}{Dataframe on which to operate. Must have column DateStamp}

\item{summarizationMethod}{String that indicates mean or sum}

\item{flipNA}{TRUE or FALSE to convert NAs to 0}
}
\description{
Summarize data frame by Month
This function summarizes all columns by week and returns a dataframe with summarized data and month dates
}
\examples{
SummarizeByMonth(df, "mean", TRUE)
}
\keyword{month}
\keyword{monthly}
\keyword{summarize}

